make Favourite button work with new table and FavouriteController

+++ database/migrations/2024_04_12_110040_add_foreign_best_answer_id_to_questions_table.php
@@ -12,13 +12,13 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('questions', function (Blueprint $table) {
-           
+            if (!Schema::hasColumn('questions', 'best_answer_id')) {
                 $table->unsignedBigInteger('best_answer_id')->nullable();
                 $table->foreign('best_answer_id')
                       ->references('id')
                       ->on('answers')
                       ->onDelete('SET NULL');
-            
+            }
         });
     }
